Zmiana sposobu sposobu tworzenia zapytan gorm.io w paginacji

+++ server/models/Animal.go
@@ -16,7 +16,7 @@ type Animal struct {
 	Adoptable     bool      `json:"adoptable"`
 	Description   string    `json:"description"`
 	Age           uint      `json:"age"`
-	Weight        uint      `json:"weight"`
+	Weight        float32   `json:"weight"`
 	Sex           string    `json:"sex"`
 	AdmissionDate time.Time `json:"admission_date"`
 	ChipNumber    string    `json:"chip_number"`
